requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
cssselect==1.2.0
Cython==3.0.11
debugpy==1.8.6
decorator==4.4.2
distro==1.9.0
dtw-python==1.5.3
einops==0.8.0
einx==0.3.0
elevenlabs==1.9.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.16.1
flatbuffers==24.3.25
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
jedi==0.19.1
jieba==0.42.1
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json-stream==2.3.2
json-stream-rs-tokenizer==0.4.26
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
llvmlite==0.43.0
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
narwhals==1.14.2
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.1
openai==1.51.2
openai-whisper==20240930
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
playsound==1.2.2
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pypinyin==0.53.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
QtPy==2.4.1
qtreload==0.1.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.2
rpds-py==0.21.0
scipy==1.14.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.40.2
superqt==0.6.7
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tinysegmenter==0.3
tldextract==5.1.3
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
websockets==13.1
whisper-timestamped==1.15.8
widgetsnbextension==4.0.13
yarl==1.14.0
yt-dlp==2024.11.18
